requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
aiohttp
httpx